Move fetching and resolution of existingClientID to SetupForm.
