+++ assets/js/modules/sign-in-with-google/components/common/ClientIDTextField.js
@@ -20,26 +20,23 @@
  * External dependencies
  */
 import classnames from 'classnames';
-import { useMount } from 'react-use';
 
 /**
  * WordPress dependencies
  */
-import { useCallback, useState } from '@wordpress/element';
+import { useCallback, useState, useEffect } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import { useSelect, useDispatch, useRegistry } from 'googlesitekit-data';
+import { useSelect, useDispatch } from 'googlesitekit-data';
 import { TextField } from 'googlesitekit-components';
 import { MODULES_SIGN_IN_WITH_GOOGLE } from '../../datastore/constants';
 import { isValidClientID } from '../../utils/validation';
 import { useDebounce } from '../../../../hooks/useDebounce';
 
-export default function ClientIDTextField() {
-	const registry = useRegistry();
-
+export default function ClientIDTextField( { existingClientID = '' } ) {
 	const clientID = useSelect( ( select ) =>
 		select( MODULES_SIGN_IN_WITH_GOOGLE ).getClientID()
 	);
@@ -52,6 +49,13 @@ export default function ClientIDTextField() {
 	const debounceSetIsValid = useDebounce( setIsValid, 500 );
 
 	const { setClientID } = useDispatch( MODULES_SIGN_IN_WITH_GOOGLE );
+
+	useEffect( () => {
+		if ( ! clientID && existingClientID ) {
+			setClientID( existingClientID );
+		}
+	}, [ clientID, setClientID, existingClientID ] );
+
 	const onChange = useCallback(
 		( { currentTarget } ) => {
 			const newValue = currentTarget.value;
@@ -65,34 +69,6 @@ export default function ClientIDTextField() {
 		},
 		[ clientID, setClientID, debounceSetIsValid ]
 	);
-
-	// Prefill the clientID field with a value from a previous module connection, if it exists.
-	useMount( async () => {
-		// Allow default `settings` and `savedSettings` to load before updating
-		// the `clientID` setting again.
-		await registry
-			.resolveSelect( MODULES_SIGN_IN_WITH_GOOGLE )
-			.getSettings();
-
-		// The clientID is fetched again as useMount does not receive the
-		// updated clientID.
-		const currentClientID = registry
-			.select( MODULES_SIGN_IN_WITH_GOOGLE )
-			.getClientID();
-
-		if (
-			currentClientID === '' &&
-			global._googlesitekitModulesData?.[ 'sign-in-with-google' ][
-				'existingClientID'
-			]
-		) {
-			setClientID(
-				global._googlesitekitModulesData[ 'sign-in-with-google' ]
-					.existingClientID
-			);
-			setIsExistingClientID( true );
-		}
-	} );
 
 	let helperText;
 	if ( ! isValid ) {
